+++ packages/blockchain-link/src/workers/solana/handlers/subscribe.ts
@@ -1,0 +1,25 @@
+import type { MessageTypes } from '@trezor/blockchain-link-types';
+import { CustomError, RESPONSES } from '@trezor/blockchain-link-types';
+
+import { subscribeAccounts } from '../subscriptions/accounts';
+import { subscribeBlock } from '../subscriptions/block';
+import type { Request } from '../types';
+
+export const subscribe = async (request: Request<MessageTypes.Subscribe>) => {
+    let response: { subscribed: boolean };
+    switch (request.payload.type) {
+        case 'block':
+            response = await subscribeBlock(request);
+            break;
+        case 'accounts':
+            response = await subscribeAccounts(request, request.payload.accounts);
+            break;
+        default:
+            throw new CustomError('worker_unknown_request', `+${request.type}`);
+    }
+
+    return {
+        type: RESPONSES.SUBSCRIBE,
+        payload: response,
+    } as const;
+};
