+++ packages/blockchain-link/src/workers/solana/handlers/unsubscribe.ts
@@ -1,0 +1,25 @@
+import type { MessageTypes } from '@trezor/blockchain-link-types';
+import { CustomError, RESPONSES } from '@trezor/blockchain-link-types';
+
+import { unsubscribeAccounts } from '../subscriptions/accounts';
+import { unsubscribeBlock } from '../subscriptions/block';
+import type { Request } from '../types';
+
+export const unsubscribe = (request: Request<MessageTypes.Unsubscribe>) => {
+    switch (request.payload.type) {
+        case 'block':
+            unsubscribeBlock(request);
+            break;
+        case 'accounts': {
+            unsubscribeAccounts(request, request.payload.accounts);
+            break;
+        }
+        default:
+            throw new CustomError('worker_unknown_request', `+${request.type}`);
+    }
+
+    return {
+        type: RESPONSES.UNSUBSCRIBE,
+        payload: { subscribed: request.state.getAccounts().length > 0 },
+    } as const;
+};
