+++ packages/blockchain-link/src/workers/solana/types.ts
@@ -1,0 +1,16 @@
+import type { TokenDetailByMint } from '@trezor/blockchain-link-types';
+import type { Signature, Slot, SolanaAPI } from '@trezor/network-solana/types';
+
+import type { ContextType } from '../baseWorker';
+
+export type Context = ContextType<SolanaAPI> & {
+    getTokenMetadata: () => Promise<TokenDetailByMint>;
+    onNetworkDisconnect: () => void;
+};
+
+export type Request<T> = T & Context;
+
+export type SignatureWithSlot = {
+    signature: Signature;
+    slot: Slot;
+};
